+++ packages/bezier-react/src/components/Tooltip/TooltipContent.tsx
@@ -82,8 +82,8 @@ const TooltipContent: React.FC<TooltipContentProps> = ({
   content,
   contentClassName,
   contentInterpolation,
+  contentWrapperStyle: givenContentWrapperStyle,
   disabled = false,
-  container: givenContainer,
   keepInContainer = false,
   placement = TooltipPosition.BottomCenter,
   tooltipContainer,
@@ -97,8 +97,6 @@ const TooltipContent: React.FC<TooltipContentProps> = ({
   const mergedRef = useMergeRefs<HTMLDivElement>(tooltipRef, forwardedRef)
   const [replacement, setReplacement] = useState(placement)
 
-  const container = givenContainer || getRootElement()
-
   const handleClickTooltip = useCallback((event: HTMLElementEventMap['click']) => {
     event.stopPropagation()
   }, [])
@@ -109,13 +107,11 @@ const TooltipContent: React.FC<TooltipContentProps> = ({
     if (!tooltipRef.current) { return }
     const newPlacement = getReplacement({
       tooltip: tooltipRef.current,
-      container,
       keepInContainer,
       placement,
     })
     setReplacement(newPlacement)
   }, [
-    container,
     keepInContainer,
     placement,
   ])
@@ -124,17 +120,21 @@ const TooltipContent: React.FC<TooltipContentProps> = ({
 
   const contentWrapperStyle = useMemo(() => {
     if (tooltipContainer) {
-      return getTooltipStyle({
-        tooltipContainer,
-        placement: replacement,
-        offset,
-        allowHover,
-      })
+      return {
+        ...givenContentWrapperStyle,
+        ...getTooltipStyle({
+          tooltipContainer,
+          placement: replacement,
+          offset,
+          allowHover,
+        }),
+      }
     }
 
     return {}
   }, [
     tooltipContainer,
+    givenContentWrapperStyle,
     replacement,
     offset,
     allowHover,
@@ -159,7 +159,7 @@ const TooltipContent: React.FC<TooltipContentProps> = ({
           </EllipsisableContent>
         </Content>
       </ContentWrapper>,
-      container,
+      getRootElement(),
     )
   )
 }