+++ packages/bezier-react/src/features/Feature.ts
@@ -1,0 +1,15 @@
+export enum FeatureType {
+  SmoothCorners = 'smooth-corners',
+}
+
+export class Feature {
+  name: FeatureType
+
+  constructor(name: FeatureType) {
+    this.name = name
+  }
+
+  async activate() {
+    return Promise.resolve(false)
+  }
+}