+++ packages/bezier-react/src/features/index.ts
@@ -1,0 +1,9 @@
+export { FeatureType } from './Feature'
+
+export {
+  FeatureProvider,
+  useFeatureFlag,
+  type FeatureFlag,
+} from './FeatureProvider'
+
+export { SmoothCornersFeature } from './SmoothCornersFeature'